dhcpv4: consolidate string options into one file.
